[map] Fetch tiles for a given bounding box or coordinate
Also introduces the Tile abstraction and the VectorTile.

When creating a Map, we now have to specify what kind of map we
are creating using Generics, like Map<VectorTile> or in the future
Map<RasterTile> or Map<GeoJSONTile>.

We can add methods, properties, etc at will to the VectorTile class.
The only thing it has to implement from the Tile interface for now
is a method to decode the data and make a resource, so we know how
to request this particular tile.