[build] Fix serialization on IL2CPP
IL2CPP will try to strip code that is not in use. For some
reason attributes of classes like GeocodeResponse are being
stripped, so JSON is not getting serialized properly.

This fix forces the Mapbox.Directions and Mapbox.Geocoding
assemblies to get preserved on the code stripping phase.

See ongoing discussion:
https://github.com/SaladLab/Json.Net.Unity3D/issues/7